Add Checkbox and Filter Functionality
I did numerous things behind the scenes as far as experimentation, but
on this commit I've added a simple checkbox signifying that a given
scripture should be read again, as well as a filter feature to choose
the "greatness" level of any given piece of scripture. I made these
changes mainly so that I could use this page to follow along with a
tutorial I'm doing on JS Fundamentals.