Add Backend Functionality For Input
I've added a function (addItem) into the script that will later be used
to trigger several other functions that take text input and add the text
input to a list using a button. I'm following a long with a tutorial at
the moment and will likely recreate this project again several times.