Begin Adding Local Storage Functionality
I've added the initital steps of local storage functionality so that
when an item is added through the "Enter Scripture" input that the item
is then set into an array titled "items" after being converted to a
string using JSON stringify method. I don't fully understand this but
I will do this more until it becomes clear.